# api/openapi.yaml
openapi: 3.1.2
info:
  title: Synodus Core API
  version: 1.0.0
  description: >-
    The reviewed v1 browser API contract. Organization, team, and task
    operations remain dark until the D06 route cutover.
...
paths:
...
  /api/v1/orgs:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      summary: List organizations visible to the user
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: Organization page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OrganizationPage'}
              example:
                items:
                  - {id: 11111111-1111-4111-8111-111111111111, name: Applied Systems Lab, lifecycle_state: active, created_at: '2026-08-27T09:00:00Z', updated_at: '2026-08-27T09:00:00Z', deleted_at: null}
                next_cursor: null
                previous_cursor: null
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    post:
      tags: [Organizations]
      operationId: createOrganization
      summary: Create an organization
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrganizationWrite'}
            example: {name: Applied Systems Lab}
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
              examples: {standard: {$ref: '#/components/examples/Organization'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      tags: [Organizations]
      operationId: getOrganization
      summary: Get an organization
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Organization
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
              examples: {standard: {$ref: '#/components/examples/Organization'}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    patch:
      tags: [Organizations]
      operationId: updateOrganization
      summary: Rename an organization
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrganizationWrite'}
            example: {name: Applied Systems Laboratory}
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
              examples: {standard: {$ref: '#/components/examples/Organization'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    delete:
      tags: [Organizations]
      operationId: deleteOrganization
      summary: Soft-delete an organization
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Organization deleted}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/restore:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    post:
      tags: [Organizations]
      operationId: restoreOrganization
      summary: Restore a soft-deleted organization
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '200':
          description: Organization restored
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
              examples: {standard: {$ref: '#/components/examples/Organization'}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/teams:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      tags: [Teams]
      operationId: listTeams
      summary: List teams in an organization
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: Team page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TeamPage'}
              example:
                items:
                  - {id: 22222222-2222-4222-8222-222222222222, name: Reproducibility Group, slug: reproducibility-group, created_at: '2026-08-27T09:00:00Z', updated_at: '2026-08-27T09:00:00Z'}
                next_cursor: null
                previous_cursor: null
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    post:
      tags: [Teams]
      operationId: createTeam
      summary: Create a team
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TeamWrite'}
            example: {name: Reproducibility Group}
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Team'}
              examples: {standard: {$ref: '#/components/examples/Team'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
...
  /api/v1/orgs/{org_id}/members:
    parameters: [{$ref: '#/components/parameters/OrganizationID'}]
    get:
      operationId: listOrganizationMembers
      security: [{sessionCookie: []}]
      responses: {'200': {description: Organization members}}
...
  /api/v1/orgs/{org_id}/teams/{team_id}/members:
    parameters:
      - {$ref: '#/components/parameters/OrganizationID'}
      - {$ref: '#/components/parameters/TeamID'}
    get:
      operationId: listTeamMembers
      security: [{sessionCookie: []}]
      responses: {'200': {description: Team members}}
    post:
      operationId: addTeamMember
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [email], properties: {email: {type: string, format: email}}}}}
      responses: {'201': {description: Team member added}}
  /api/v1/orgs/{org_id}/teams/{team_id}/tasks:
    parameters:
      - {$ref: '#/components/parameters/OrganizationID'}
      - {$ref: '#/components/parameters/TeamID'}
    get:
      tags: [Tasks]
      operationId: listTasks
      summary: List tasks in a team
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: Task page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskPage'}
              example:
                items:
                  - {id: 33333333-3333-4333-8333-333333333333, description: Reproduce benchmark results, status: todo, created_at: '2026-08-27T09:00:00Z', updated_at: '2026-08-27T09:00:00Z'}
                next_cursor: null
                previous_cursor: null
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    post:
      tags: [Tasks]
      operationId: createTask
      summary: Create a task
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TaskWrite'}
            example: {description: Reproduce benchmark results, status: todo}
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Task'}
              examples: {standard: {$ref: '#/components/examples/Task'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/v1/orgs/{org_id}/teams/{team_id}/tasks/{task_id}:
    parameters:
      - {$ref: '#/components/parameters/OrganizationID'}
      - {$ref: '#/components/parameters/TeamID'}
      - {$ref: '#/components/parameters/TaskID'}
    patch:
      tags: [Tasks]
      operationId: updateTask
      summary: Update a task
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TaskWrite'}
            example: {description: Reproduce benchmark results, status: in_progress}
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Task'}
              examples: {standard: {$ref: '#/components/examples/Task'}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    delete:
      tags: [Tasks]
      operationId: deleteTask
      summary: Delete a task
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Task deleted}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalFailure'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-synodus_session
      description: Production browser session cookie.
...
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 512}
...
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
    Cursor:
      name: cursor
      in: query
      schema: {type: string, minLength: 1, maxLength: 512}
    OrganizationID:
      name: org_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PublicID'}
    TeamID:
      name: team_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PublicID'}
    TaskID:
      name: task_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PublicID'}
...
  responses:
    InvalidRequest: {$ref: '#/components/responses/InvalidRequestResponse'}
    Unauthenticated: {$ref: '#/components/responses/UnauthenticatedResponse'}
    Forbidden: {$ref: '#/components/responses/ForbiddenResponse'}
    NotFound: {$ref: '#/components/responses/NotFoundResponse'}
    Conflict: {$ref: '#/components/responses/ConflictResponse'}
    PreconditionFailed: {$ref: '#/components/responses/PreconditionFailedResponse'}
    RateLimited: {$ref: '#/components/responses/RateLimitedResponse'}
    InternalFailure: {$ref: '#/components/responses/InternalFailureResponse'}
    DependencyUnavailable: {$ref: '#/components/responses/DependencyUnavailableResponse'}
    InvalidRequestResponse:
      description: The request is invalid
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          examples: {standard: {$ref: '#/components/examples/InvalidRequestProblem'}}
    UnauthenticatedResponse:
      description: Authentication is required
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          examples: {standard: {$ref: '#/components/examples/UnauthenticatedProblem'}}
...
  schemas:
    PublicID: {type: string, format: uuid}
...
    Organization:
      type: object
      additionalProperties: false
      required: [id, name, lifecycle_state, created_at, updated_at]
      properties:
        id: {$ref: '#/components/schemas/PublicID'}
        name: {type: string, minLength: 1, maxLength: 200}
        lifecycle_state: {type: string, enum: [provisioning, active, deleting, deleted, restoring, failed]}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
        deleted_at: {type: [string, 'null'], format: date-time}
    OrganizationWrite:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 200}
    Team:
      type: object
      additionalProperties: false
      required: [id, name, slug, created_at, updated_at]
      properties:
        id: {$ref: '#/components/schemas/PublicID'}
        name: {type: string, minLength: 1, maxLength: 200}
        slug: {type: string, minLength: 1, maxLength: 100, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
    TeamWrite:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 200}
    Task:
      type: object
      additionalProperties: false
      required: [id, description, status, created_at, updated_at]
      properties:
        id: {$ref: '#/components/schemas/PublicID'}
        description: {type: string, minLength: 1, maxLength: 4096}
        status: {type: string, enum: [todo, in_progress, done]}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
    TaskWrite:
      type: object
      additionalProperties: false
      required: [description, status]
      properties:
        description: {type: string, minLength: 1, maxLength: 4096}
        status: {type: string, enum: [todo, in_progress, done]}
...
    OrganizationPage:
      allOf:
        - {$ref: '#/components/schemas/PageLinks'}
        - type: object
          required: [items]
          properties:
            items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/Organization'}}
    TeamPage:
      allOf:
        - {$ref: '#/components/schemas/PageLinks'}
        - type: object
          required: [items]
          properties:
            items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/Team'}}
    TaskPage:
      allOf:
        - {$ref: '#/components/schemas/PageLinks'}
        - type: object
          required: [items]
          properties:
            items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/Task'}}
    PageLinks:
      type: object
      required: [next_cursor, previous_cursor]
      properties:
        next_cursor: {$ref: '#/components/schemas/CursorValue'}
        previous_cursor: {$ref: '#/components/schemas/CursorValue'}
...
  examples:
...
    Organization:
      value: {id: 11111111-1111-4111-8111-111111111111, name: Applied Systems Lab, lifecycle_state: active, created_at: '2026-08-27T09:00:00Z', updated_at: '2026-08-27T09:00:00Z', deleted_at: null}
    Team:
      value: {id: 22222222-2222-4222-8222-222222222222, name: Reproducibility Group, slug: reproducibility-group, created_at: '2026-08-27T09:00:00Z', updated_at: '2026-08-27T09:00:00Z'}
    Task:
      value: {id: 33333333-3333-4333-8333-333333333333, description: Reproduce benchmark results, status: todo, created_at: '2026-08-27T09:00:00Z', updated_at: '2026-08-27T09:00:00Z'}
...